Isolate yubikey-related capabilities to the plugins/yubikey subdirectory
