+++ docs/installation.rst
@@ -35,6 +35,7 @@ Add the following apps to the ``INSTALLED_APPS``:
         'django_otp.plugins.otp_totp',
         'two_factor',
         'two_factor.plugins.phonenumber',  # <- if you want phone number capability.
+        'two_factor.plugins.yubikey',  # <- for yubikey capability.
     )
 
 Add the ``django-otp`` middleware to your ``MIDDLEWARE``. Make sure
