+++ two_factor/plugins/yubikey/forms.py
@@ -1,0 +1,16 @@
+from django import forms
+from django.utils.translation import gettext_lazy as _
+
+from two_factor.forms import DeviceValidationForm
+
+
+class YubiKeyDeviceForm(DeviceValidationForm):
+    token = forms.CharField(label=_("YubiKey"), widget=forms.PasswordInput())
+
+    error_messages = {
+        'invalid_token': _("The YubiKey could not be verified."),
+    }
+
+    def clean_token(self):
+        self.device.public_id = self.cleaned_data['token'][:-32]
+        return super().clean_token()
